% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbExtract.R
\name{dbExtract}
\alias{dbExtract}
\title{dbEtract main function}
\usage{
dbExtract(inputFile = "../raw/inputs/dbInputs.csv",
  catFile = "../raw/inputs/categories.csv", output = "dbExtractOutput")
}
\arguments{
\item{inputFile}{inputFile filename, needs to be in the working directory}

\item{dirPath}{the path to the databases (see input file)}

\item{startAt}{at which line to start in input}

\item{append}{= F,}
}
\description{
This function is designed to explore databases and summarize
the spatial and temporal coverage of pre-selected varaibles (need to fill the input.xls file)
}
\examples{
dbExtract ()
}
\keyword{cats}
